fix toFloat

+++ ex01/main.cpp
@@ -6,7 +6,7 @@
 /*   By: antoine <antoine@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/01 21:31:50 by antoine           #+#    #+#             */
-/*   Updated: 2022/08/03 18:46:03 by antoine          ###   ########.fr       */
+/*   Updated: 2022/08/04 13:48:14 by aalleon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,15 +29,15 @@ int	main()
 	Fixed	f(99);
 
 	std::cout << 42 << " -> " << d.toInt() << std::endl;
-	std::cout << -42 << " -> " << e.toInt() << std::endl;
+	std::cout << -42 << " -> " << e << std::endl;
 	std::cout << 99 << " -> " << f.toInt() << std::endl;
 
 	Fixed	g(42.12f);
 	Fixed	h(1.111111111f);
 	Fixed	i(-42.12454f);
 
-	std::cout << 42.12f << "->" << g.toFloat() << std::endl;
-	std::cout << 1.111111111f << "->" << h.toFloat() << std::endl;
-	std::cout << -42.12454f << "->" << i.toFloat() << std::endl;
+	std::cout << 42.12f << "->" << g << std::endl;
+	std::cout << 1.111111111f << "->" << h << std::endl;
+	std::cout << -42.12454f << "->" << i << std::endl;
 	return (0);
 }
